# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/locataire/edit/{id}": {
            "put": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Update a utilisateurs",
                "operationId": "49d41f059ceaf23bda2091e53c902751",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the utilisateurs",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "type_user_id": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "addresse": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "statut": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/locataire/create": {
            "post": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Create a new  locataire",
                "operationId": "79f49b6f8a9ae14c0aee1e99b4297ffc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "nom",
                                    "telephone",
                                    "addresse"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "sexe": {
                                        "type": "string"
                                    },
                                    "login": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "statut": {
                                        "type": "string"
                                    },
                                    "type_user_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "utilisateurs created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/locataire/show/{id}": {
            "get": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Get a specific resource",
                "operationId": "e874812f2db4b8ec16ac41921e34e594",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": " id de utilisateurs",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
        "/api/v1/locataire/delete/{id}": {
            "delete": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Delete a locataire",
                "operationId": "f0eed3437bcaadbfaadd222c9497be81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    }
}